requirements.txt
python-whois
pyfiglet
colorama
builtwith